quarry: Created 0000-new-quarry-block
Introducing **Quarry of Creation**.

It's a project scaffolder, similar to Python's Cookiecutter
(with a pinch of inspiration from Symfony Flex recipes).

In the **"Quarry"** (repository), we find a list of **"Blocks"**, each
including an `install.php` script that contains a list of executable steps.

These steps will use templates to create or modify small parts of a project,
to help write and evolve it.

The plan is to be able to have blocks that allow:

* bootstraping generic project documentation
  (LICENSE, README, CONTRIBUTING, VERSIONING, installation, usage, etc)
* bootstraping generic project tooling
  (makefile, docker, etc)
* bootstraping generic PHP projects, and PHP tooling configuration
  (PHPUnit, PHPStan, PHP CS Fixer with rules, etc)
* and more

> _Worth Noting_: Quarry of Creation is built using itself.
> **Continuous Dogfooding** ensures its iterative refinement and relevance.

So in that regard, the first Block provided allows... Bootstrapping Blocks.